fix mock server bindings
