+++ src/it/java/uk/gov/caz/psr/SuccessPaymentsJourneyTestIT.java
@@ -33,7 +33,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockserver.integration.ClientAndServer;
-import org.mockserver.model.Header;
 import org.mockserver.model.HttpRequest;
 import org.mockserver.model.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -96,7 +95,6 @@ public class SuccessPaymentsJourneyTestIT extends VccsCallsIT {
   private SecretsManagerInitialisation secretsManagerInitialisation;
 
   private ClientAndServer mockServer;
-  private ClientAndServer vccsMockServer;
 
   @Test
   public void testPaymentJourneys() {
@@ -244,7 +242,6 @@ public class SuccessPaymentsJourneyTestIT extends VccsCallsIT {
   @BeforeEach
   public void startMockServer() {
     mockServer = startClientAndServer(1080);
-    vccsMockServer = startClientAndServer(1090);
     RestAssured.port = randomServerPort;
     RestAssured.baseURI = "http://localhost";
     RestAssured.basePath = "/v1/payments";
@@ -265,7 +262,6 @@ public class SuccessPaymentsJourneyTestIT extends VccsCallsIT {
   @AfterEach
   public void stopMockServer() {
     mockServer.stop();
-    vccsMockServer.stop();
   }
 
   @AfterEach
@@ -564,17 +560,6 @@ public class SuccessPaymentsJourneyTestIT extends VccsCallsIT {
             .withStatusCode(HttpStatus.OK.value())
             .withHeader("Content-type", MediaType.APPLICATION_JSON.toString())
             .withBody(readFile("get-payment-response.json")));
-  }
-
-  public void mockVccsResponse() {
-    vccsMockServer
-        .when(HttpRequest.request()
-            .withPath("/v1/compliance-checker/clean-air-zones")
-            .withMethod("GET"))
-        .respond(HttpResponse.response()
-            .withStatusCode(200)
-            .withHeaders(new Header("Content-Type", "application/json; charset=utf-8"))
-            .withBody(readFile("get-clean-air-zones.json")));
   }
 
   /// ----- utility methods
