+++ src/it/java/uk/gov/caz/psr/VccsCallsIT.java
@@ -1,7 +1,8 @@
 package uk.gov.caz.psr;
 
 import static org.mockserver.integration.ClientAndServer.startClientAndServer;
-
+import static org.mockserver.matchers.Times.exactly;
+import java.nio.file.Files;
 import com.google.common.base.Charsets;
 import com.google.common.io.Resources;
 import lombok.SneakyThrows;
@@ -11,6 +12,7 @@ import org.mockserver.integration.ClientAndServer;
 import org.mockserver.model.Header;
 import org.mockserver.model.HttpRequest;
 import org.mockserver.model.HttpResponse;
+import org.springframework.util.ResourceUtils;
 
 public class VccsCallsIT {
 
@@ -36,10 +38,41 @@ public class VccsCallsIT {
             .withHeaders(new Header("Content-Type", "application/json; charset=utf-8"))
             .withBody(readFile("get-clean-air-zones.json")));
   }
+  
+  
+  public void mockVccsComplianceCall(String vrn) {
+    vccsMockServer
+        .when(requestGet("/v1/compliance-checker/vehicles/" + vrn + "/compliance"),
+            exactly(1))
+        .respond(response("vehicle-compliance-response.json"));
+  }
 
   @SneakyThrows
   private String readFile(String filename) {
     return Resources.toString(Resources.getResource("data/external/" + filename),
         Charsets.UTF_8);
+  }
+  
+  @SneakyThrows
+  public static String readJson(String file) {
+    return new String(
+        Files.readAllBytes(ResourceUtils.getFile("classpath:data/json/response/" + file).toPath()));
+  }
+
+  public static HttpResponse response(String responseFile) {
+    return HttpResponse.response()
+        .withStatusCode(200)
+        .withHeaders(new Header("Content-Type", "application/json; charset=utf-8"))
+        .withBody(readJson(responseFile));
+  }
+
+  public static HttpRequest requestGet(String url) {
+    return prepareRequest(url, "GET");
+  }
+
+  public static HttpRequest prepareRequest(String url, String method) {
+    return HttpRequest.request()
+        .withPath(url)
+        .withMethod(method);
   }
 }
